+++ cong-nghe-web-dattestbr/src/main/java/com/HTTN/thitn/controller/SubjectController.java
@@ -33,7 +33,7 @@ public class SubjectController {
 
     @PutMapping("/teacher/subjects/{id}")
 
-    public ResponseEntity<SubjectResponse> updateSubject(@PathVariable Integer id, @Valid @RequestBody SubjectRequest request) {
+    public ResponseEntity<SubjectResponse> updateSubject(@PathVariable Long id, @Valid @RequestBody SubjectRequest request) {
         Subject subject = new Subject();
         subject.setName(request.getName());
         subject.setDescription(request.getDescription());
@@ -43,7 +43,7 @@ public class SubjectController {
 
 
     @GetMapping("/teacher/subjects/{id}")
-    public ResponseEntity<SubjectResponse> getSubjectById(@PathVariable Integer id) {
+    public ResponseEntity<SubjectResponse> getSubjectById(@PathVariable Long id) {
         Subject subject = subjectService.getSubjectById(id);
         if (subject == null) {
             // Trả về mã lỗi 404 nếu không tìm thấy subject
@@ -62,7 +62,7 @@ public class SubjectController {
     }
 // giáo viên thêm hs vào
     @PostMapping("/teacher/subjects/{subjectId}/students/{studentId}")
-    public ResponseEntity<?> addStudentToSubject(@PathVariable Integer subjectId, @PathVariable Long studentId) {
+    public ResponseEntity<?> addStudentToSubject(@PathVariable Long subjectId, @PathVariable Long studentId) {
         boolean added = subjectService.addStudentToSubject(subjectId, studentId);
         if (added) {
             return ResponseEntity.ok().body("Học sinh đã được thêm vào môn học.");
@@ -71,7 +71,7 @@ public class SubjectController {
         }
     }
     @GetMapping("/teacher/subjects/{subjectId}/students")
-    public ResponseEntity<?> getStudentsInSubject(@PathVariable Integer subjectId) {
+    public ResponseEntity<?> getStudentsInSubject(@PathVariable Long subjectId) {
         List<User> students = subjectService.getStudentsInSubject(subjectId);
         if (students == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Không tìm thấy môn học.");
